requirements.txt
absl-py==1.4.0
antlr4-python3-runtime==4.8
appdirs==1.4.4
astunparse==1.6.3
attrs==22.2.0
autoflake==1.7.8
awscli==1.27.94
black==21.4b0
boto3==1.26.94
botocore==1.29.94
cachetools==5.3.0
certifi @ file:///croot/certifi_1671487769961/work/certifi
chardet==3.0.4
charset-normalizer==3.1.0
click==8.1.3
cloudpickle==2.2.1
cmake==3.26.0
colorama==0.4.4
coloredlogs==15.0.1
decorator==5.1.1
dill==0.3.6
dm-tree==0.1.8
docutils==0.16
exceptiongroup==1.1.1
filelock==3.10.0
fire==0.5.0
flake8==3.8.1
flatbuffers==2.0.7
flowvision==0.1.0
ftfy==6.1.1
gast==0.4.0
google-auth==2.16.2
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
graphviz==0.8.4
grpcio==1.51.3
h5py==3.8.0
huggingface-hub==0.13.2
humanfriendly==10.0
hydra-core==1.1.2
idna==2.8
importlib-metadata==6.0.0
importlib-resources==5.2.3
iniconfig==2.0.0
isort==5.10.1
jieba==0.42.1
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.2.0
keras==2.11.0
libclang==15.0.6.1
lit==15.0.7
Markdown==3.4.1
markdown-it-py==2.2.0
MarkupSafe==2.1.2
mccabe==0.6.1
mdurl==0.1.2
mpmath==1.3.0
mxnet-mkl==1.6.0
mypy-extensions==1.0.0
networkx==3.0
nltk==3.8.1
numpy==1.23.4
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
oauthlib==3.2.2
omegaconf==2.1.0
onefx==0.0.3
onnx==1.13.1
onnx-simplifier==0.4.17
onnx-tf==1.10.0
onnxruntime==1.14.1
onnxsim==0.4.19              
opencv-python==4.7.0.72
opt-einsum==3.3.0
packaging==23.0
pathspec==0.11.1
Pillow==9.4.0
pluggy==1.0.0
portalocker==2.7.0
protobuf==3.20.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pybind11==2.10.4
pycodestyle==2.6.0
pyflakes==2.2.0
Pygments==2.14.0
pytest==7.2.2
python-dateutil==2.8.2
PyYAML==6.0
regex==2017.4.5
requests==2.21.0
requests-oauthlib==1.3.1
rich==13.3.2
rsa==4.9
s3transfer==0.6.0
scipy==1.10.1
sentencepiece==0.1.97
six==1.16.0
sympy==1.11.1
tabulate==0.9.0
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorboardX==2.6
tensorflow-addons==0.19.0
tensorflow-cpu==2.11.0
tensorflow-estimator==2.11.0
tensorflow-io-gcs-filesystem==0.31.0
tensorflow-probability==0.19.0
termcolor==2.2.0
tf2onnx==1.13.0
tokenizers==0.13.2
toml==0.10.2
tomli==2.0.1
torch==2.0.0
torch-summary==1.4.5
torchvision==0.15.1
tqdm==4.65.0
transformers==4.27.1
triton==2.0.0
typeguard==3.0.1
typing_extensions==4.5.0
urllib3==1.26.15
wcwidth==0.2.6
Werkzeug==2.2.3
wget==3.2
wrapt==1.15.0
zipp==3.15.0
